src/third_party/blink/web_tests/virtual/web-components-v0-disabled/fast/dom/shadow/shadow-root-blur-expected.txt

CONSOLE ERROR: line 87: Uncaught TypeError: element.createShadowRoot is not a function
This tests the blur() on on shadow host should work when a shadow host contains a focused element in its shadow DOM subtrees property (bug 81102)

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


FAIL successfullyParsed should be true. Was false.

TEST COMPLETE
